Work in progress. a13r1.
This is alpha 13, release 1.
Now I try to translate and checking with the game which use quite a lot
of times, I will do this when available, so please be patient. A lot of
Defs are still in Deutsch and still a bit works in Keyed.
Thank you for understanding.